+++ client/app/components/postlist/postlist.component.ts
@@ -1,5 +1,5 @@
 import { Component, OnInit, ChangeDetectorRef } from '@angular/core';
-import { BlogService } from 'client/app/services/blogservice/blog.service';
+import { PostManagerService } from 'client/app/services/postmanager/postmanager.service';
 import { BlogPost } from 'client/app/models/blogpost';
 
 @Component({
@@ -9,7 +9,7 @@ import { BlogPost } from 'client/app/models/blogpost';
 })
 export class PostlistComponent implements OnInit {
 
-  constructor(private postService:BlogService, private cdr : ChangeDetectorRef) { 
+  constructor(private postService:PostManagerService, private cdr : ChangeDetectorRef) { 
 
     this.postService.setNewPostAction(() => {
       this.newPost();
