+++ client/app/components/toolbar/toolbar.component.ts
@@ -1,7 +1,7 @@
 import { Component, OnInit, ChangeDetectorRef } from '@angular/core';
-import { BlogService } from 'client/app/services/blogservice/blog.service';
+import { PostManagerService } from 'client/app/services/postmanager/postmanager.service';
 import { Blog } from 'client/app/models/blog';
-import { NavigationService } from 'client/app/services/navigationservice/navigation.service';
+import { AppManagerService } from 'client/app/services/appmanager/appmanager.service';
 
 @Component({
   selector: 'app-toolbar',
@@ -10,58 +10,63 @@ import { NavigationService } from 'client/app/services/navigationservice/navigat
 })
 export class ToolbarComponent implements OnInit {
 
-  constructor(private blogservice: BlogService, private navService: NavigationService, private cdr : ChangeDetectorRef) { }
+  constructor(private blogservice: PostManagerService, private navService: AppManagerService, private cdr : ChangeDetectorRef) { }
 
   ngOnInit() {
 
     this.toggleEditor();
-    this.blogservice.fetchConfigurations();
-    this.blogservice.listenForMenuInvocation();
+    this.navService.fetchConfigurations();
+    this.navService.listenForMenuInvocation();
 
-    this.blogservice.menuListener.on('save', (args) => {
+    this.navService.getMenuEventEmitter().on('save', (args) => {
       this.savePost();
-      this.blogservice.updateListener.emit('postUpdated');
+      this.blogservice.postUpdateListener.emit('postUpdated');
     });
 
-    this.blogservice.menuListener.on('new', (args) => {
+    this.navService.getMenuEventEmitter().on('new', (args) => {
       this.newPost();
-      this.blogservice.updateListener.emit('postUpdated');
+      this.blogservice.postUpdateListener.emit('postUpdated');
     });
 
-    this.blogservice.menuListener.on('htmlEditor', (args) => {
+    this.navService.getMenuEventEmitter().on('htmlEditor', (args) => {
       this.toggleEditor();
-      this.blogservice.updateListener.emit("postUpdated");
+      this.blogservice.postUpdateListener.emit("postUpdated");
     });
 
     // listens for publish to blog menu event
-    this.blogservice.menuListener.on('publishToBlog', (args) => {
+    this.navService.getMenuEventEmitter().on('publishToBlog', (args) => {
       var blogObj = args.blog;
       var blog = new Blog(blogObj.name, blogObj.url, blogObj.id);
       this.publishBlog(blog, false);
     });
 
     // listens for draft to blog menu event
-    this.blogservice.menuListener.on('draftToBlog', (args) => {
+    this.navService.getMenuEventEmitter().on('draftToBlog', (args) => {
       var blogObj = args.blog;
       var blog = new Blog(blogObj.name, blogObj.url, blogObj.id);
       this.publishBlog(blog, true);
     });
 
     // listens for sidebar collapse/expand
-    this.blogservice.menuListener.on('toggleSidePanel', (args) => {
+    this.navService.getMenuEventEmitter().on('toggleSidePanel', (args) => {
       this.navService.setPanelHidden(!args.viewSideBar);
       this.cdr.detectChanges();
     })
 
     // listens for any updates to posts
-    this.blogservice.updateListener.on('postUpdated', () => {
+    this.blogservice.postUpdateListener.on('postUpdated', () => {
+      this.cdr.detectChanges();
+    })
+
+
+    this.navService.getUIEventEmitter().on('uiUpdated', (args) => {
       this.cdr.detectChanges();
     })
 
   }
 
   getBlogs():Blog[] {
-    return this.blogservice.getBlogs();
+    return this.navService.getBlogsList();
   }
 
   // pubish a blog
@@ -81,13 +86,12 @@ export class ToolbarComponent implements OnInit {
 
   // switches between HTML editor and RT editor
   toggleEditor() {
-    this.blogservice.setHTMLEditor(!this.blogservice.isHTMLEditor());
-    this.cdr.detectChanges();
+    this.navService.setHTMLEditor(!this.navService.isHTMLEditor());
   }
 
   // returns if the current editor is HTML editor
   isHTMLEditor() {
-    return this.blogservice.isHTMLEditor();
+    return this.navService.isHTMLEditor();
   }
 
 }
